Add cryptowatch Summary support to Framework

+++ bitcoin-de-sdk-value/src/main/kotlin/de/nykon/bitcoin/sdk/value/cryptowatch/summary/Change.kt
@@ -1,6 +1,8 @@
 package de.nykon.bitcoin.sdk.value.cryptowatch.summary
 
+import java.math.BigDecimal
+
 data class Change(
-    val absolute: Double,
-    val percentage: Double
+        val absolute: BigDecimal,
+        val percentage: BigDecimal
 )
