+++ bitcoin-de-sdk-value/src/main/kotlin/de/nykon/bitcoin/sdk/value/cryptowatch/summary/Price.kt
@@ -1,8 +1,10 @@
 package de.nykon.bitcoin.sdk.value.cryptowatch.summary
 
+import java.math.BigDecimal
+
 data class Price(
-    val change: Change,
-    val high: Double,
-    val last: Double,
-    val low: Int
+        val change: Change,
+        val high: BigDecimal,
+        val last: BigDecimal,
+        val low: BigDecimal
 )
