+++ bitcoin-de-sdk-value/src/main/kotlin/de/nykon/bitcoin/sdk/value/cryptowatch/summary/Result.kt
@@ -1,7 +1,9 @@
 package de.nykon.bitcoin.sdk.value.cryptowatch.summary
 
+import java.math.BigDecimal
+
 data class Result(
     val price: Price,
-    val volume: Double,
-    val volumeQuote: Double
+    val volume: BigDecimal,
+    val volumeQuote: BigDecimal
 )
